+++ src/test/com/yandex/app/managers/InMemoryTaskManagerTest.java
@@ -1,4 +1,4 @@
-package com.yandex.app.managers;
+package test.com.yandex.app.managers;
 
 import com.yandex.app.model.Epic;
 import com.yandex.app.model.Subtask;
